Expose 'RepositoryProvider' in the analysis result ('Repository.provider').
