+++ lib/src/model.dart
@@ -377,20 +377,38 @@ class AnalysisResult {
 ///   - must not have a `publish_to` key.
 @JsonSerializable()
 class Repository {
+  final String provider;
   final String baseUrl;
   final String? branch;
   final String? packagePath;
 
   Repository({
+    String? provider,
     required this.baseUrl,
     this.branch,
     this.packagePath,
-  });
+  }) : provider = provider ?? RepositoryProvider.unknown;
 
   factory Repository.fromJson(Map<String, dynamic> json) =>
       _$RepositoryFromJson(json);
 
   Map<String, dynamic> toJson() => _$RepositoryToJson(this);
+}
+
+/// The identifier of repository provider, which could influence how relative URLs are resolved.
+abstract class RepositoryProvider {
+  /// GitHub
+  static const github = 'github';
+
+  /// GitLab (cloud or self-hosted)
+  static const gitlab = 'gitlab';
+
+  /// Unable to identify.
+  static const unknown = 'unknown';
+
+  /// Whether the provider follows the GitHub URL conventions.
+  static bool isGitHubCompatible(String? provider) =>
+      provider == github || provider == gitlab;
 }
 
 @JsonSerializable()
