+++ test/goldens/end2end/bulma_min-0.7.4.json
@@ -100,6 +100,7 @@
     "repositoryUrl": "https://github.com/agilord/bulma_min",
     "issueTrackerUrl": "https://github.com/agilord/bulma_min/issues",
     "repository": {
+      "provider": "github",
       "baseUrl": "https://github.com/agilord/bulma_min",
       "branch": "master"
     }
