+++ test/goldens/end2end/nsd_android-1.2.0.json
@@ -107,6 +107,7 @@
     "repositoryUrl": "https://github.com/sebastianhaberey/nsd/tree/main/nsd_android",
     "issueTrackerUrl": "https://github.com/sebastianhaberey/nsd/issues",
     "repository": {
+      "provider": "github",
       "baseUrl": "https://github.com/sebastianhaberey/nsd",
       "branch": "main",
       "packagePath": "nsd_android"
