+++ test/goldens/end2end/skiplist-0.1.0.json
@@ -95,6 +95,7 @@
     "repositoryUrl": "https://github.com/stevenroose/dart-skiplist",
     "issueTrackerUrl": "https://github.com/stevenroose/dart-skiplist/issues",
     "repository": {
+      "provider": "github",
       "baseUrl": "https://github.com/stevenroose/dart-skiplist",
       "branch": "master"
     }
