+++ test/goldens/end2end/webdriver-3.0.0.json
@@ -113,6 +113,7 @@
     "repositoryUrl": "https://github.com/google/webdriver.dart",
     "issueTrackerUrl": "https://github.com/google/webdriver.dart/issues",
     "repository": {
+      "provider": "github",
       "baseUrl": "https://github.com/google/webdriver.dart",
       "branch": "master"
     }
